improvement: speechactclassifier extracted to own module
