+++ opentutor_classifier/svm/__init__.py
@@ -36,10 +36,9 @@ from opentutor_classifier import (
     load_yaml,
     QuestionConfig,
     ExpectationFeatures,
-    SpeechActClassifierResult,
     TrainingResult,
 )
-
+from opentutor_classifier.speechact import SpeechActClassifier
 from opentutor_classifier.stopwords import STOPWORDS
 
 WORD2VEC_MODELS: Dict[str, Word2VecKeyedVectors] = {}
@@ -483,26 +482,6 @@ class SVMAnswerClassifierTraining:
         return self.accuracy
 
 
-class SpeechActClassifier:
-    def __init__(self):
-        self.meta_regex = r"\b(idk|belie\w*|don\w*|comprehend\w*|confuse\w*|guess\w*|(?<=n[o']t)\s?\b(know\w*|underst\w*|follow\w*|recogniz\w*|sure\w*|get)\b|messed|no\s?(idea|clue)|lost|forg[eo]t|need\s?help|imagined?|interpret(ed)?|(seen?|saw)|suppos(ed)?)\b"
-        self.profanity = r"\b(\w*fuck\w*|ass|ass(hole|wipe|wad|kisser)|hell|shit|piss\w*|cock|cock(sucker|head|eater)|douche\w*|bitch\w*|retard[ed]|midget\w*|dyke|fag|faggot|cunt\w*|\w*nigg\w*|tranny|slut\w*|cum[bucket]|dick\w*|pussy\w*|dildo\w*|idiot\w*|(hate you)|stupid\w*)\b"
-
-    def check_meta_cognitive(self, result):
-        input_sentence = result.input.input_sentence
-        if re.search(self.meta_regex, input_sentence, re.IGNORECASE):
-            return SpeechActClassifierResult(evaluation="Good", score=1)
-        else:
-            return SpeechActClassifierResult(evaluation="Bad", score=0)
-
-    def check_profanity(self, result):
-        input_sentence = result.input.input_sentence
-        if re.search(self.profanity, input_sentence, re.IGNORECASE):
-            return SpeechActClassifierResult(evaluation="Good", score=1)
-        else:
-            return SpeechActClassifierResult(evaluation="Bad", score=0)
-
-
 class SVMAnswerClassifier:
     def __init__(self, model_root="models", shared_root="shared"):
         self.model_root = model_root
@@ -510,7 +489,7 @@ class SVMAnswerClassifier:
         self.model_obj = SVMExpectationClassifier()
         self._word2vec = None
         self._instance_models = None
-        self.speech_act_obj = SpeechActClassifier()
+        self.speech_act_classifier = SpeechActClassifier()
 
     def instance_models(self) -> InstanceModels:
         if not self._instance_models:
@@ -566,10 +545,12 @@ class SVMAnswerClassifier:
         word2vec = self.find_word2vec()
         index2word = set(word2vec.index2word)
 
-        result.speech_acts["metacognitive"] = self.speech_act_obj.check_meta_cognitive(
+        result.speech_acts[
+            "metacognitive"
+        ] = self.speech_act_classifier.check_meta_cognitive(result)
+        result.speech_acts["profanity"] = self.speech_act_classifier.check_profanity(
             result
         )
-        result.speech_acts["profanity"] = self.speech_act_obj.check_profanity(result)
         print("resul = ", result)
         if answer.config_data:
             conf = answer.config_data
